+++ src/models/HiddenDefaultTemplateFormContext.php
@@ -1,0 +1,10 @@
+<?php
+namespace verbb\formie\models;
+
+use craft\base\Model;
+
+class HiddenDefaultTemplateFormContext extends Model
+{
+    public string $handle = '';
+    public string $title = '';
+}