+++ src/models/HiddenDefaultTemplateSiteContext.php
@@ -1,0 +1,11 @@
+<?php
+namespace verbb\formie\models;
+
+use craft\base\Model;
+
+class HiddenDefaultTemplateSiteContext extends Model
+{
+    public int $id = 0;
+    public string $handle = '';
+    public string $name = '';
+}